Moved abilities to classes

+++ Test/PlayerRenderComponent.cpp
@@ -53,8 +53,8 @@ void PlayerRenderComponent::Update(){
 	// ugly way of seeing if this is the local player
 	if (!gameObjectRef->GetComponent(COMPONENT_INPUT))
 		return;
-	crosshairRef->posX = inputMan->getMouseX() - crosshairRef->getWidth()/2;
-	crosshairRef->posY = inputMan->getMouseY() - crosshairRef->getHeight()/2;
+	crosshairRef->posX = (float)(inputMan->getMouseX() - crosshairRef->getWidth()/2);
+	crosshairRef->posY = (float)(inputMan->getMouseY() - crosshairRef->getHeight() / 2);
 
 }
